requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
aiohttp==3.9.1
python-telegram-bot==20.6
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
pyyaml==6.0.1
httpx==0.25.1
websockets==12.0
python-dotenv==1.0.0
